requirements.txt
fastapi[all]==0.75.0
uvicorn[standard]==0.15.0
SQLAlchemy==1.4.34
pydantic==1.9.0
black==22.3.0
redis==4.2.2
python-decouple==3.6
python-dotenv==0.20.0 
psycopg2-binary==2.9.3

requests==2.27.1